Berni branch 2 updated ryans code

+++ ryan.py
@@ -1,10 +1,16 @@
+# Step one, checking if code is existant and downloaded if not
 import urllib.request
-#import request module (on PSL)
+import os
 
 url = 'https://s3.amazonaws.com/tcmg476/http_access_log'
-#specify url
 filename = 'log.txt'
-#name dowloaded file
 
-#download file
-urllib.request.urlretrieve(url, filename)
+# Check if the file already exists
+if os.path.exists(filename):
+    print(f"The file '{filename}' already exists.")
+else:
+    print(f"Downloading '{filename}'...")
+    urllib.request.urlretrieve(url, filename)
+    print(f"'{filename}' has been downloaded.")
+
+# rest of the code will go after this
